+++ src/index.ts
@@ -121,6 +121,7 @@ export * from "./lib/logger/types/log-level";
 
 export * from "./lib/types/entity";
 export * from "./lib/types/class-type";
+export * from "./lib/types/extra-metadata";
 
 /**
  * FindOptions