+++ CHANGELOG.md
@@ -30,6 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Ex: `const connection = await new Connection().load().connect()`
 - [Internal] Update `options` to completely remove `entities` and `entitiesDir`
 - Now loads both `entities` and `entitiesDir` options, so both can be used at the same time
+- [BC] Update extra metadata to allow infinite extra metadata
+  - Now the extra metadata is one unique object, and it's properties are used to get the types
 
 ### Fixed